% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abracadabra.R
\name{adjust_text}
\alias{adjust_text}
\title{Remove last characters from text and adjust it}
\usage{
adjust_text(text, flag = " ")
}
\arguments{
\item{text}{string; text to be cropped.}

\item{flag}{character; everything after the last time this character appears will be removed.}
}
\value{
string; text cropped and adjusted.
}
\description{
Cuts everything on a string after last time a flag character is present.
}
