% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abracadabra.R
\name{is.int}
\alias{is.int}
\title{Verify if a variable is as integer number}
\usage{
is.int(x)
}
\arguments{
\item{x}{numeric.}
}
\value{
boolean;
\itemize{
\item \code{TRUE}  if \code{x} is an integer number,
\item \code{FALSE} if it is not.
}
}
\description{
Verify if a variable is as integer number
}
