% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lupicinio.R
\name{Lupicinio}
\alias{Lupicinio}
\title{Get a song lyric by Lupicínio every day}
\usage{
Lupicinio(max.char = NULL, flag = " ", seed = NULL)
}
\arguments{
\item{max.char}{integer; maximum number of characters to be returned; if NULL (DEFAULT) the text is not cropped.}

\item{flag}{character; if max.char is used, everything after the last time this character appears will be removed; the DEFAULT is a single space " ".}

\item{seed}{numeric; a seed for the draw; if NULL (DEFAULT) no seed is set.}
}
\value{
list; contains author, title, text, and source.
}
\description{
Draw a random song lyric by Brazilian musician Lupicínio Rodrigues from
<www.letras.mus.br/> based on an seed.
}
\examples{
Lupicinio(max.char = 1000)
# should return the first 1000 ish characters of today's Lupicínio song lyric.

}
